WIP | trying to get example working
